Make `simd_inner_length` generic fallback less specific

+++ base/simdloop.jl
@@ -39,15 +39,19 @@ check_body!(x) = true
 # @simd splits a for loop into two loops: an outer scalar loop and
 # an inner loop marked with :simdloop. The simd_... functions define
 # the splitting.
+# Custom iterators that do not support random access cannot support
+# vectorization. In order to be compatible with `@simd` annotated loops,
+#they should override `simd_inner_length(v::MyIter, j) = 1`,
+#`simd_outer_range(v::MyIter) = v`, and `simd_index(v::MyIter, j, i) = j`.
 
 # Get range for outer loop.
 simd_outer_range(r) = 0:0
 
 # Get trip count for inner loop.
-@inline simd_inner_length(r,j::Int) = Base.length(r)
+@inline simd_inner_length(r, j) = Base.length(r)
 
 # Construct user-level element from original range, outer loop index j, and inner loop index i.
-@inline simd_index(r,j::Int,i) = (@inbounds ret = r[i+firstindex(r)]; ret)
+@inline simd_index(r, j, i) = (@inbounds ret = r[i+firstindex(r)]; ret)
 
 # Compile Expr x in context of @simd.
 function compile(x, ivdep)
